# Simplify OrderPreservingInterner allocation strategy ~97% faster (#2677)