# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #2677

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

Not only was the previous allocation strategy quite hard to follow, when interning sorted data it would end up comparing the value against at least `1/4` of the already interned values. This was sub-optimal.

Taking a step back the original approach of leaving spaces was to mitigate the impact of unsorted data, as the design evolved `OrderPreservingInterner` now always sorts the dictionary prior to interning. This dramatically reduces the likelihood of out-of-order inserts, and consequently diminishes the value of this mitigation.

A simpler, in-order allocation strategy is both faster and easier to understand

```
Benchmarking convert_columns 4096 string_dictionary(10, 0): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.6s, enable flat sampling, or reduce sample count to 60.
convert_columns 4096 string_dictionary(10, 0)
                        time:   [1.1147 ms 1.1156 ms 1.1169 ms]
                        change: [-96.615% -96.612% -96.607%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

Benchmarking convert_columns 4096 string_dictionary(30, 0): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.8s, enable flat sampling, or reduce sample count to 60.
convert_columns 4096 string_dictionary(30, 0)
                        time:   [1.1510 ms 1.1514 ms 1.1519 ms]
                        change: [-96.623% -96.621% -96.619%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Benchmarking convert_columns 4096 string_dictionary(100, 0): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.1s, enable flat sampling, or reduce sample count to 60.
convert_columns 4096 string_dictionary(100, 0)
                        time:   [690.29 µs 692.36 µs 695.23 µs]
                        change: [-97.493% -97.335% -97.211%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  1 (1.00%) low mild
  19 (19.00%) high severe

convert_columns 4096 string_dictionary(100, 0.5)
                        time:   [316.77 µs 316.89 µs 317.00 µs]
                        change: [-95.710% -95.708% -95.705%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

Changes the allocation strategy for OrderPreservingInterner to be simpler to understand and faster.

# Are there any user-facing changes?

No

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->